pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["murtazox04 <murtazox04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.1"
pydantic-settings = "^2.3.4"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
